fixed: filter signal inversion, PAL/NTSC timing precision, ADSR thresholds, licensing inconsistency, additional delay-bug plan-B case

+++ src/env.h
@@ -24,7 +24,7 @@ void envSetPsidMode(uint8_t m);
 
 uint8_t envIsFilePSID();	// original file setting
 
-int8_t envFPS();
+double envFPS();
 int32_t envClockRate();
 
 uint8_t envIsSID6581();
